+++ lib/enumerable.rb
